move defaults filler script to processing
