+++ flex-tasks-processing/package.yaml
@@ -4,6 +4,7 @@ extra-source-files: [ ]
 dependencies:
   - base
   - text
+  - shakespeare
 default-extensions:
   - TupleSections
   - DuplicateRecordFields
@@ -18,6 +19,7 @@ library:
   source-dirs: src
   exposed-modules:
     - FlexTask.Processing.Text
+    - FlexTask.Processing.JavaScript
   ghc-options:
     - -Wall
     - -Widentities
